Issue #31: Organize folders and files
Organizing documents
