+++ SeededImageSegmentationProject.pro
@@ -1,0 +1,32 @@
+######################################################################
+# Automatically generated by qmake (2.01a) lun nov 9 23:39:34 2015
+######################################################################
+
+TEMPLATE = app
+TARGET = 
+DEPENDPATH += . Common Communication Exceptions GUI MathTools
+INCLUDEPATH += . GUI Communication MathTools Common Exceptions
+CONFIG += link_pkgconfig
+PKGCONFIG += opencv eigen3
+
+# Input
+HEADERS += Common/neighbourhood.h \
+           Common/neighbourhoodfactory.h \
+           Communication/segmentationeventhandler.h \
+           Exceptions/exception.h \
+           Exceptions/mathexception.h \
+           Exceptions/userinputexception.h \
+           GUI/displaywindow.h \
+           GUI/mainwindow.h \
+           MathTools/seededsegmentation.h
+FORMS += GUI/mainwindow.ui
+SOURCES += main.cpp \
+           Common/neighbourhood.cpp \
+           Common/neighbourhoodfactory.cpp \
+           Communication/segmentationeventhandler.cpp \
+           Exceptions/exception.cpp \
+           Exceptions/mathexception.cpp \
+           Exceptions/userinputexception.cpp \
+           GUI/displaywindow.cpp \
+           GUI/mainwindow.cpp \
+           MathTools/seededsegmentation.cpp
